Split type-aware lint rules into modules
